+++ packages/grafana-data/src/index.ts
@@ -4,3 +4,4 @@ export * from './vector';
 export * from './dataframe';
 export * from './transformations';
 export * from './datetime';
+export * from './text';
