+++ packages/grafana-data/src/utils/index.ts
@@ -1,6 +1,4 @@
-export * from './string';
 export * from './registry';
-export * from './markdown';
 export * from './deprecationWarning';
 export * from './csv';
 export * from './fieldReducer';
@@ -9,6 +7,5 @@ export * from './labels';
 export * from './labels';
 export * from './object';
 export * from './thresholds';
-export * from './text';
 
 export { getMappedValue } from './valueMappings';
